Re-add the big malloc mode, use it if possible.
The big 512M allocation is tried first and if it fails, try the smaller
one. We encode the selected mode (full or compressed) using the LSB of
the base_mem pointer.
This should allow systems with enough virtual memory to use the full mem
base mode, while still allowing other systems to use the compressed mem
base mode.